requirements.txt
alembic==0.8.9
cffi==1.9.1
click==6.6
cryptography==1.7.1
ecdsa==0.13
enum34==1.1.6
Flask==0.12
Flask-Cache==0.13.1
Flask-Login==0.4.0
Flask-Migrate==2.0.2
Flask-Script==2.0.5
Flask-SQLAlchemy==2.1
Flask-WTF==0.13.1
gunicorn==19.6.0
idna==2.2
ipaddress==1.0.17
itsdangerous==0.24
Jinja2==2.8
libvirt-python==1.3.1
Mako==1.0.6
MarkupSafe==0.23
MySQL-python==1.2.3
paramiko==2.1.1
pkg-resources==0.0.0
pyasn1==0.1.9
pycparser==2.17
pycrypto==2.6.1
python-editor==1.0.3
six==1.10.0
SQLAlchemy==1.1.4
Werkzeug==0.11.11
WTForms==2.1
